完善登录功能，登录拦截，密码加密等等

+++ src/main/java/org/employeems/entity/permissions/User.java
@@ -36,7 +36,7 @@ public class User {
     /**
      * 是否启用（0-禁用，1-启用）
      */
-    private Byte enabled;
+    private Integer enabled;
 
     /**
      * 创建时间
